Feature/PROTO-33 improve sonar score and coverag

+++ code_coverage.sh
@@ -1,4 +1,4 @@
-#! /bin/sh
+#! /bin/bash
 
 #
 # Copyright (C) 2020 Embedded AMS B.V. - All Rights Reserved
@@ -30,19 +30,32 @@
 #   the Netherlands
 #
 
-./build/test/test_EmbeddedProto
+./build/test/test_EmbeddedProto --gtest_output="xml:build/test/test_details.xml"
 
-# rm -rf ./code_coverage_report/*
-# lcov --directory ./build/test --capture --output-file ./code_coverage_report/total_code_coverage.info -rc lcov_branch_coverage=1
-# lcov --remove ./code_coverage_report/total_code_coverage.info $PWD'/external/googletest/*' '/usr/include/*' -o ./code_coverage_report/filtered_code_coverage.info
-# genhtml ./code_coverage_report/filtered_code_coverage.info --branch-coverage --output-directory ./code_coverage_report
 
-rm -rf ./code_coverage_report/*
-mkdir -p code_coverage_report
-cd code_coverage_report
+if [ $# -eq 0 ]; then
+  # No arguments provided, execte the analyisis for sonar qube.
+  rm -rf ./code_coverage_report/*
+  mkdir -p code_coverage_report
+  cd code_coverage_report
 
-# Run gcov on the static source files.
-gcov ../test/*.cpp --object-directory ../build/test/CMakeFiles/test_EmbeddedProto.dir/test/ 
-gcov ../src/*.cpp --object-directory ../build/test/CMakeFiles/test_EmbeddedProto.dir/src/
+  # Run gcov on the static source files.
+  gcov ../test/*.cpp --object-directory ../build/test/CMakeFiles/test_EmbeddedProto.dir/test/ 
+  gcov ../src/*.cpp --object-directory ../build/test/CMakeFiles/test_EmbeddedProto.dir/src/
 
-cd -
+  cd -
+
+else
+# Pass commands
+case "$1" in
+  -l|--local)
+    # Run the code coverage for local machine analysis
+    rm -rf ./code_coverage_report/*
+    lcov --directory ./build/test --capture --output-file ./code_coverage_report/total_code_coverage.info -rc lcov_branch_coverage=1
+    lcov --remove ./code_coverage_report/total_code_coverage.info $PWD'/external/googletest/*' '/usr/include/*' -o ./code_coverage_report/filtered_code_coverage.info
+    genhtml ./code_coverage_report/filtered_code_coverage.info --branch-coverage --output-directory ./code_coverage_report
+    ;;
+  *)
+    ;;
+esac
+fi
